Bug fix for BTDF sample point plotting in debug version (pabopto2rad)

+++ src/cv/pabopto2bsdf.c
@@ -238,7 +238,7 @@ main(int argc, char *argv[])
 		if (dsf_grid[i][j].sum.n > 0) {
 			ovec_from_pos(dir, i, j);
 			bsdf = dsf_grid[i][j].sum.v /
-				(dsf_grid[i][j].sum.n*output_orient*dir[2]);
+			   ((double)dsf_grid[i][j].sum.n*output_orient*dir[2]);
 			if (bsdf <= bsdf_min*.6)
 				continue;
 			bsdf = log(bsdf + 1e-5) - min_log;
